Feature/refactor

+++ src/handler/transaction.handler.ts
@@ -1,22 +1,21 @@
-import { Action, Command, Hears, On, Update } from 'nestjs-telegraf';
+import { Action, On, Update } from 'nestjs-telegraf';
 import { TransactionService } from '../service';
 import { Logger } from '@nestjs/common';
 import {
   BALANCE_MESSAGE,
   ENTER_EXPENSE_MESSAGE,
   ENTER_INCOME_MESSAGE,
-  ERROR_MESSAGE,
   getBalanceMessage,
   INVALID_DATA_MESSAGE,
-  INVALID_TRANSACTION_NAME_MESSAGE,
   SELECT_TRANSACTION_MESSAGE,
   TRANSACTION_DELETED_MESSAGE,
 } from '../constants/messages';
-import { actionButtonsTransaction } from '../battons/app.buttons';
+import { actionButtonsTransaction, backTranButton } from '../battons/app.buttons';
 import { CustomCallbackQuery, IContext } from '../interface/context.interface';
 import { MyMessage } from '../interface/my-message.interface';
 import { BalanceService } from '../service';
 import { TransactionType } from '../shemas/enum/transactionType.enam';
+import { checkAndUpdateLastBotMessage } from '../utils/botUtils';
 
 @Update()
 export class TransactionHandler {
@@ -26,57 +25,85 @@ export class TransactionHandler {
     private readonly balanceService: BalanceService,
   ) {}
 
-  @Command('transactions')
-  @Hears(/Transactions 💸|Транзакції 💸/)
+  @Action('transactions')
   async aboutCommand(ctx: IContext) {
-    await ctx.deleteMessage();
+    if (await checkAndUpdateLastBotMessage(ctx)) {
+      return;
+    }
     delete ctx.session.type;
-    this.logger.log('Транзакция command executed');
-    await ctx.replyWithHTML(
+    this.logger.log('transactions command executed');
+    await ctx.telegram.editMessageText(
+      ctx.from.id,
+      ctx.session.lastBotMessage,
+      null,
       SELECT_TRANSACTION_MESSAGE[ctx.session.language || 'ua'],
       actionButtonsTransaction(ctx.session.language),
     );
   }
-  @Action('Приход')
+  @Action('income')
   async incomeCommand(ctx: IContext) {
+    if (await checkAndUpdateLastBotMessage(ctx)) {
+      return;
+    }
     ctx.session.type = 'income';
-    await ctx.deleteMessage();
-    await ctx.replyWithHTML(ENTER_INCOME_MESSAGE[ctx.session.language || 'ua']);
+    await ctx.telegram.editMessageText(
+      ctx.from.id,
+      ctx.session.lastBotMessage,
+      null,
+      ENTER_INCOME_MESSAGE[ctx.session.language || 'ua'],
+      backTranButton(ctx.session.language || 'ua'),
+    );
 
     this.logger.log('Приход command executed');
   }
 
-  @Action('Расход')
+  @Action('expense')
   async expenseCommand(ctx: IContext) {
+    if (await checkAndUpdateLastBotMessage(ctx)) {
+      return;
+    }
     ctx.session.type = 'expense';
-    await ctx.deleteMessage();
-    await ctx.replyWithHTML(ENTER_EXPENSE_MESSAGE[ctx.session.language || 'ua']);
+    await ctx.telegram.editMessageText(
+      ctx.from.id,
+      ctx.session.lastBotMessage,
+      null,
+      ENTER_EXPENSE_MESSAGE[ctx.session.language || 'ua'],
+      backTranButton(ctx.session.language || 'ua'),
+    );
 
     this.logger.log('Расход command executed');
   }
-  @Action('Удаление последних️')
+  @Action('delete_last')
   async deleteLastCommand(ctx: IContext) {
-    const userId = ctx.from.id;
+    if (await checkAndUpdateLastBotMessage(ctx)) {
+      return;
+    }
     ctx.session.type = 'delete';
-    await this.transactionService.showLastNTransactionsWithDeleteOption(userId, 20, ctx.session.language || 'ua');
+    const count = 20;
+    await this.transactionService.showLastNTransactionsWithDeleteOption(ctx, count);
   }
   @Action(/delete_(.+)/)
   async handleCallbackQuery(ctx: IContext) {
+    if (await checkAndUpdateLastBotMessage(ctx)) {
+      return;
+    }
     try {
       if (ctx.session.type !== 'delete') {
         return;
       }
-      await ctx.deleteMessage();
-
       const customCallbackQuery: CustomCallbackQuery = ctx.callbackQuery as CustomCallbackQuery;
-
       if (customCallbackQuery && 'data' in customCallbackQuery) {
         const callbackData = customCallbackQuery.data;
-        const userId = ctx.from.id;
         if (callbackData.startsWith('delete_')) {
           const transactionIdToDelete = callbackData.replace('delete_', '');
-          await this.transactionService.deleteTransactionById(userId, transactionIdToDelete);
-          await ctx.reply(TRANSACTION_DELETED_MESSAGE[ctx.session.language || 'ua']);
+          await this.transactionService.deleteTransactionById(ctx, transactionIdToDelete);
+          await ctx.telegram.editMessageText(
+            ctx.from.id,
+            ctx.session.lastBotMessage,
+            null,
+            TRANSACTION_DELETED_MESSAGE[ctx.session.language || 'ua'],
+            backTranButton(ctx.session.language || 'ua'),
+          );
           delete ctx.session.type;
         }
       } else {
@@ -90,32 +117,44 @@ export class TransactionHandler {
 
   @On('text')
   async textCommand(ctx: IContext) {
+    if (await checkAndUpdateLastBotMessage(ctx)) {
+      return;
+    }
     if (ctx.session.type !== 'income' && ctx.session.type !== 'expense') {
       return;
     }
+
     const message = ctx.message as MyMessage;
     const userId = ctx.from.id;
     const text = message.text;
     const transactions = text.split(',').map((t) => t.trim());
 
+    let errorMessageSent = false;
+
     for (const transaction of transactions) {
       const regex = /^([a-zA-Zа-яА-ЯіІ]+(?:\s+[a-zA-Zа-яА-ЯіІ]+)?)\s+([\d.]+)$/;
       const matches = transaction.match(regex);
+
       if (!matches) {
-        await ctx.reply(INVALID_DATA_MESSAGE[ctx.session.language || 'ua']);
-        return;
+        errorMessageSent = true;
+        continue;
       }
+
       const transactionName = matches[1].trim().toLowerCase();
       const amount = Number(matches[2]);
+
       if (!transactionName || isNaN(amount) || amount <= 0) {
-        await ctx.reply(INVALID_DATA_MESSAGE[ctx.session.language || 'ua']);
-        return;
+        errorMessageSent = true;
+        continue;
       }
+
       const words = transactionName.split(' ');
+
       if (words.length > 2) {
-        await ctx.reply(INVALID_TRANSACTION_NAME_MESSAGE[ctx.session.language || 'ua']);
-        return;
+        errorMessageSent = true;
+        continue;
       }
+
       const transactionType = ctx.session.type === 'income' ? TransactionType.INCOME : TransactionType.EXPENSE;
 
       try {
@@ -126,16 +165,52 @@ export class TransactionHandler {
           amount,
         });
         await this.balanceService.updateBalance(userId, amount, transactionType);
-        await ctx.reply(BALANCE_MESSAGE[ctx.session.language || 'ua']);
-        const balance = await this.balanceService.getBalance(userId, ctx.session.group);
-        const balanceMessage = getBalanceMessage(balance, ctx.session.language || 'ua', ctx.session.currency || 'UAH');
-        await ctx.replyWithHTML(balanceMessage);
-        this.logger.log('textCommand executed');
       } catch (error) {
-        // this.logger.error('Error in textCommand:', error);
-        await ctx.reply(ERROR_MESSAGE[ctx.session.language || 'ua']);
+        this.logger.error('Error creating transaction:', error);
+        errorMessageSent = true;
       }
     }
+    if (errorMessageSent) {
+      await ctx.deleteMessage();
+      await ctx.telegram.editMessageText(
+        ctx.from.id,
+        ctx.session.lastBotMessage,
+        null,
+        INVALID_DATA_MESSAGE[ctx.session.language || 'ua'],
+        backTranButton(ctx.session.language || 'ua'),
+      );
+    } else {
+      await ctx.deleteMessage();
+      const balance = await this.balanceService.getBalance(userId, ctx.session.group);
+      const balanceMessage = getBalanceMessage(balance, ctx.session.language || 'ua', ctx.session.currency || 'UAH');
+      await ctx.telegram.editMessageText(
+        ctx.from.id,
+        ctx.session.lastBotMessage,
+        null,
+        `${BALANCE_MESSAGE[ctx.session.language]}\n${balanceMessage}`,
+        {
+          parse_mode: 'HTML',
+          reply_markup: backTranButton(ctx.session.language || 'ua').reply_markup,
+        },
+      );
+      this.logger.log('textCommand executed');
+    }
+
     delete ctx.session.type;
+  }
+
+  @Action('backT')
+  async backT(ctx: IContext) {
+    if (await checkAndUpdateLastBotMessage(ctx)) {
+      return;
+    }
+    delete ctx.session.type;
+    await ctx.telegram.editMessageText(
+      ctx.from.id,
+      ctx.session.lastBotMessage,
+      null,
+      SELECT_TRANSACTION_MESSAGE[ctx.session.language || 'ua'],
+      actionButtonsTransaction(ctx.session.language || 'ua'),
+    );
   }
 }
